feat(4-4): 接入 Milvus VectorStore
- 为 rag-cli 增加 store 参数，支持 memory 和 milvus 存储切换
- 将 VectorStore 接口改为 context-aware，并统一 Add/Search 调用签名
- 将 embedding/vector 类型调整为 float32，适配 Milvus FloatVector
- 将 Milvus 存储实现迁移到 rag 包，补充 collection schema、插入、搜索和结果解析逻辑
- 初始化 Milvus collection 时增加存在性检查，并统一执行 LoadCollection + Await
- 增加 Milvus 配置项、EMBEDDING_DIM 示例配置和雪花 ID 生成工具
- 适配 RAG 相关测试用例

+++ go-ai-agent/internal/rag/chunker.go
@@ -36,9 +36,9 @@ func ChunkDocument(doc *Document, size, overlap int) ([]*Chunk, error) {
 		chunk := &Chunk{
 			Title:           doc.Title,
 			SourceFile:      doc.SourcePath,
-			RuneStartOffset: start,
-			RuneEndOffset:   end,
-			ChunkIndex:      chunkIdx,
+			RuneStartOffset: int64(start),
+			RuneEndOffset:   int64(end),
+			ChunkIndex:      int64(chunkIdx),
 			CreatedAt:       timestamp,
 			UpdatedAt:       timestamp,
 		}
