+++ go-ai-agent/internal/rag/embedding.go
@@ -19,8 +19,8 @@ func NewEmbeddingClient(url, model string) Embedder {
 // Embed 调用 embedding 服务为多段 chunk 文本生成向量。
 // 输入: `ctx` 是请求上下文, `chunks` 是待生成 embedding 的文本列表。
 // 输出: 返回与输入文本一一对应的向量列表; HTTP 调用或响应解析失败时返回错误。
-// 示例: `embedder.Embed(ctx, []string{"RAG 文档问答"})` -> 返回 `[][]float64`。
-func (ec *embedderClient) Embed(ctx context.Context, chunks []string) ([][]float64, error) {
+// 示例: `embedder.Embed(ctx, []string{"RAG 文档问答"})` -> 返回 `[][]float32`。
+func (ec *embedderClient) Embed(ctx context.Context, chunks []string) ([][]float32, error) {
 	requestBody := map[string]any{
 		"model": ec.model,
 		"input": chunks,
