+++ go-ai-agent/internal/rag/embedding_test.go
@@ -13,7 +13,7 @@ import (
 // TestEmbeddingClientEmbedsLoadedWorkNoteChunks 测试使用真实 testdata 完成文档加载、chunk 切分和 embedding 请求。
 // 输入: 从 `testdata/documents/work_notes_May/五月/第三周` 加载 markdown 文档, 并把前几个 chunk 发送给模拟 Ollama 服务。
 // 输出: 返回的 embedding 数量应与输入 chunk 数量一致, 每个 embedding 应有固定维度。
-// 示例: `NewTriliumDocumentLoader(...).Load(...) -> ChunkDocument(...) -> Embed(ctx, texts)` -> 返回 `[][]float64`。
+// 示例: `NewTriliumDocumentLoader(...).Load(...) -> ChunkDocument(...) -> Embed(ctx, texts)` -> 返回 `[][]float32`。
 func TestEmbeddingClientEmbedsLoadedWorkNoteChunks(t *testing.T) {
 	documentsDir := filepath.Clean(filepath.Join("..", "..", "testdata", "documents", "work_notes_May", "五月", "第三周"))
 	loader := NewTriliumDocumentLoader(map[string]bool{".md": true}, 0)
@@ -54,9 +54,9 @@ func TestEmbeddingClientEmbedsLoadedWorkNoteChunks(t *testing.T) {
 			}
 		}
 
-		embeddings := make([][]float64, len(req.Input))
+		embeddings := make([][]float32, len(req.Input))
 		for i := range req.Input {
-			embeddings[i] = []float64{float64(i), float64(len([]rune(req.Input[i]))), 1}
+			embeddings[i] = []float32{float32(i), float32(len([]rune(req.Input[i]))), 1}
 		}
 		w.Header().Set("Content-Type", "application/json")
 		if err := json.NewEncoder(w).Encode(EmbedResp{EmbeddingsData: embeddings}); err != nil {
